Convert specialistRole in search results
The specialist role value in the serach results was a shortened camel
case version of itself. As it was shortened, the camel case filter could
not be used and the actual value must be fetched from the content.

As the search-api results don't include the framework, we're assuming
that specialist roles are the same for DOS1 and 2. This will need
looking at in the future.