Accept `documents` as well as `services` in SearchResults
In an upcoming change to the search-api ([see this PR]
(https://github.com/alphagov/digitalmarketplace-search-api/pull/124/))
we're going to be returning data structure with a top level `documents`
key instead of `services`. To allow this PR to go out without breaking
the buyer frontend, we need to be able to accept both keys for a while.