Remove temporary create user redirect
This functionality is now found at a new URL in the user FE. This
route was just a temporary redirect for people who may have been
using the original links in their emails (which were valid for
a week). Now this time has passed, we no longer need to maintain
this redirect.