Build category tree with cleaned request args
When fetching briefs from search-api for the opportunity listing
page, we never want to return briefs with the `withdrawn` status. This
means that if a request is made without the `status` filter (which would
return all briefs regardless of status), we update the request args to
include all status _except_ withdrawn.

This wasn't being added to the aggregations request to the search-api,
instead it was just using the request args, which was throwing the facet
counts off by including `withdrawn` briefs. By passing in the cleaned or
updated request args (updated args have already been cleaned), we fix
this.