Remove check for 'services' key in search results
The search-api is now only returning 'documents' so we can remove the
check for the 'services' key.

This also fixes a bug when zero search results are returned.

If there are zero results found, the response is an empty list. Using
the `get` method on a dict returns `None` instead of the empty list.
Which broke stuff.