Remove unused `link-button` functionality in shim
We are not using link button components in our applications.
On top of that, the `shimLinkWithButtonRole` has changed due to
version bumping and it no longer even accepts a parameters in it's
init method.

We have decided to keep this functionality in our tookit even though
we have no button roles in our templates at the moment or in the
frontend toolkit. This is because it can be there in case we do
add something in the future without needing to remember about this
accessibility shim which could easily be missed at code review.

https://github.com/alphagov/govuk_frontend_toolkit/commit/e882086f8e1ed00ee86c7b276b477b498ca6035a#diff-0afd84124334f16359259ae2ca718990