search page: add aria-live attributes
we don't want the whole loaded results section re-read on reloading, instead
just a shortened "N results found" summary.

the element referenced by an `aria-controls=` seems to need to be persistent,
meaning we probably can't use a fragment of the existing search summary text,
as it gets removed & replaced wholesale when a new set of results is fetched.

instead, we return the shortened summary as its own html fragment under the
summary-accessible-hint key in the live results request. this html fragment
gets placed inside a dedicated (and more importantly, persistent)
`search-summary-accessible-hint-wrapper` element, which itself can be the
target of the `aria-controls=`