Set custom config var for absolute URLs on /404 page
Using the `request.url_root` to generate absolute URLs doesn't work because the proxied requests for the /404 page have `assets.digitalmarketplace` as the request URL.

By adding a custom config variable to let the app know which URL links should be relative to means we can patch this in for requests that come in from a different domain (e.g. for the custom 404 page).