Refactor test_direct_award (API client patching part 2)
This makes use of the previous commit where we adjusted a
helper function to take the client as an argument (a pattern
we follow elsewhere) rather than importing it to the helper module directly,
which simplifies our test patching.

DownloadResultsView has the data_api_client as a class attribute
so some tests access it via self.view.data_api_client.