Instrument GA tracking to capture clicks made on downloads and external links
A lot of GA tracking comes 'out of the box' but a notable exception is tracking to capture clicks on downloads and external links. Additional code is required in order to capture these.

Currently, we are tracking some downloads and external links on the DM site (those that were specifically required for previous mission work).

The proposal is to have all downloads and external links tracked automatically across the whole site, similar to how GOV.UK have done.

The advantage of doing this is that we have key data available straight away at the start of mission teams rather than having to prioritise the work, use up dev resource and then wait for enough data to populate before being able to carry out analysis.

---> includes making sure all links on the homepage are tracked, so we can measure which links are used the most and least.

Ticket: https://trello.com/c/i912cDBe/110-instrument-ga-tracking-to-capture-clicks-made-on-downloads-and-external-links