Load all brief_search_filters manifests
 ## Summary
If we only load the brief_search_filters manifest for the latest live
DOS framework when the app starts, then as soon as we switch DOS3 on,
the search functionality will break due to not having the correct
manifest loaded.

By loading all DOS<x> brief_search_filters manifests (where we can),
this risk is minimised.