Consolidate validation mastheads
 ## Summary
Update utils to pull in `get_errors_from_wtform` (step 1 in
consolidating validation mastheads to a central template). Requires
updates to pages that use WTForms.

One notable view has been excluded - the save search page. This uses a
blend of WTForms and custom HTML to display a special kind of radio form
with some hidden questions. A tech debt ticket exists to tackle this
form but I don't have time to do it at the moment, so needs to be
prioritised for later.

 ## Ticket
https://trello.com/c/RSYEM3FL/448