Change the layout of the search page
Ticket: [trello]

Our designers @kubabartwicki and Karl Chillmaid have changed the layout
of the services search page to try and encourage users to notice and use
the 'Save your search' feature.

This PR implements those changes, which are described in detail [in the ticket][trello].
In summary, the search summary and save button are moved to the top of
the page, and the styling of the text for both is tweaked.

This commit also touches the briefs search page slightly, reducing the
margin between the page heading and subheading.

[trello]: https://trello.com/c/4ETdzfOe/