Load `services_search_filters` manifest for all frameworks
This manifest is used as part of `SearchMeta`, which can still be used
for an expired framework. By not loading it saved search pages for
expired frameworks threw a `ContentNotFound` error.

In all other places this manifest is only used with the latest live
framework, so loading for all should be fine.