task list: add new view, pre_project_task_list, and adapt views leading into direct-award buyer journey
we now show this, initial, version of the task list when entering the
journey and the existing lot selection view is renamed to `choose_lot`
and placed at a new url. the "Start a new search" action button here now
leads to that page.

this also introduces views that are under the "direct award" url path but
are not login-restricted, so a new `direct_award_public` blueprint is
introduced for this purpose.

the template is the existing "view project" template, which ironically
seems to render relatively well if called without a `project`.