Add project from project fixture to project list fixture
This is needed to ensure that tests of the save-search page continue to
pass in future.

Currently the form does not validate whether the save_search_selection
is a project that was in the list displayed to the user. If this
changes, then without this commit the tests will fail because the project
fixture POSTed to save-search was not in the project list fixture.