Use latest live framework for opportunity csv dwonload
The link to download the opportunity list csv was hardcoded to DOS2.
This could be confusing.

The CSV actually is agnostic of framework iteration so it doesn't really
matter where its served from, but keeping it in sync with the framework
is more consistent.

This will require an update to the nightly Jenkins job that generates
and uploads the file. Coming soon.