content loader: make thread-safe through use of context-local deep copies
this implementation is a little more involved than the one in the supplier
frontend, required because the buyer frontend needs information from the
application object and the api before it can populate its content loader.
hence this is performed quite late in the startup process. this should all
be fine as long as no code needs to access the methods of the
content loader before the application is initialized, though in such a case
the attempt is made to fail obviously and loudly.