Replace govuk_template template with govuk-frontend template
Switches base page template over to govuk-frontend template and replaces
govuk_template styles with govuk-frontend styles.

Remove govuk_template compatibility flag and import
digitalmarketplace-govuk-frontend stylesheet.

Currently digitalmarketplace-govuk-frontend is needed for the header
component but is likely to be needed for other components.

As we have not migrated over completely to govuk-frontend. I have
set `$govuk-global-styles` variable to true. This variable is used by
sass compiler to generate globale styles for links and paragraphs.
This will add the proper link colour to links and font size to paragraghs
Ideally all elements that need to be styled should have a class name.

We also need to update tests that were looking for the id of the main
wrapper element from govuk_template.