Restore phase tag to header on homepage
We want to keep the homepage much as it looks now; however that means we
can't use dmHeader/govukHeader macro, as that doesn't support having a
phase tag in the product name.

Instead we copy the HTML from [`govuk-frontend/components/header`][1]
into the template for the homepage so that we can put the phase tag
where we want it to be.

[1]: https://github.com/alphagov/govuk-frontend/blob/v2.13.0/src/components/header/template.njk