Add basic styles from govuk_template
We are still using some components from
digitalmarketplace-frontend-toolkit and govuk-frontend-toolkit that rely
on styles from govuk_template. Until these components can be removed or
replaced we need to have the CSS available for those styles.

This commit adds a subset of the styles included in govuk_template so
that the components still look correct.

Co-authored-by: Kat Stevens <katherine.stevens@digital.cabinet-office.gov.uk>
Co-authored-by: Alistair Laing <alistair.laing@digital.cabinet-office.gov.uk>