Decouple home page messages from temporary-message
Given the home page messages only ever use the heading and
messages options, we can remove the use of the toolkit
macro and just have the logic in the index.html template. Also
renames the div to 'temporary-framework-message' to make it
clearer what's going to be displayed in this box.