Add error styling to date inputs
We don't currently have the granularity to check which fields are
causing an error, so we highlight each field.

This isn't great accessibility. For example, a user might wonder why
a perfectly valid day value is causing an error, when in fact the error
is being caused by an invalid year value.

We should review this as part of looking into the granularity of our date
errors.