Remove use of instruction list template
Rewrites `app/templates/direct-award/view_page.html` to use HTML and if
statements, instead of the rather difficult to read instruction list
structure.

We wrap the whole thing in an ordered list and give it the
`app-task-list` class to make testing easier and to allow it to be made
more of a task list in future, but we don't actually bring in any of the
task list styles.