Move pageTitle to base_page
We want to preface titles with "Error: " if there's an error on the page. We can set a page_name on the
child page which will be used in the base_page block, or we can override the block entirely (for backwards
compatibility).

Also in this commit: Removing the unecessary wrapping divs - these classes aren't used anywhere.