Simplify the search-summary style
In order to make use of fewer custom styles, we can rejig the search summary to use
`<strong>` instead of `<em>` so that its styling is handled by govuk styles. We can
then style the element that wraps the search summary.

I've kept `app-search-summary__count` as a custom class, since we style it differently
in different places:

* On the two search pages and the "save your search" and "save a new search" pages,
it's 48px.
* On the "view project" page, it's 19px.

This change should allow us to delete `search_overview.scss` once we've removed the
instruction list there, and remove the need to add custom style overrides for the
search summary.