Add wrap arg to SearchSummary.markup()
We want to be able to wrap the search summary in a `<p>` tag with appropriate classes.

This change bakes that into the SearchSummary class.

It's possible that in the future we might want to generify this to allow us to pass anything
as a wrapper, but that's not how we use it right now, hence this approach.