Replace _brief_page.scss with _stat_block.scss
Our stat blocks are slightly misaligned with the tables that display on the opportunities pages with
them, so we've temporarily overriden them to have a smaller width. This is an issue that's been brought
up with the Design System team, so may get investigated in the future. Alternatively, we could rethink
how we display these stats.