Show 404s for deleted services
In order to support a supplier's request to be removed completely
from G-Cloud we are adding a new service status - `deleted`.

The URL for a service with `deleted` status should return a 404.
This is different from the behaviour of `disabled` status, which
still shows the service information at that URL but with a
banner stating the service is unavaible.

https://trello.com/c/qqL3sqWV/2182-remove-supplier-92526-from-g-cloud