Cache framework description
We call this function very frequently - once per page for many pages. It will only ever return one of three outputs: the G-Cloud description, the DOS description, or an empty string. The description changes very infrequently. So we can improve performance significantly by caching the output of this function.

This will be useful because a scraper currently tries to scrape 1500-2000 supplier pages very quickly once every 2 hours. We want to improve the performance of the scraped pages so that we can restore the rate limits to their old higher value.

From my testing, this reduces the time taken to respond to `/g-cloud/supplier/<supplier_id>` by 32% (from 111 ms to 76 ms).