Update accessibility statement
CCS have taken the decision not to do any further accessibility work on the DMP. This means we will not continue upgrading the components, simplifying the content or the journeys. It also means we will no longer fix any further audit findings or enforce suppliers to provide accessible documents for G-Cloud 13.
CCS have agreed to  accept any risk involved with WCAG non-compliance.