Only test interaction with are_new_frameworks_live
dmutils already has extensive tests that check that are_new_frameworks_live is behaving correctly at the correct times. So we only need to test that we're using it correctly. Remove redundant testing of behaviour at specific times. This should mean that we won't need to change the tests when we add the G13 banner.

Remove freezegun, since we no longer need it.